Got a weird timer bug when showing Story Royale on another PC.
Trying to fix it when I'm getting access to it. Still strange because it worked on my laptop which had similar specs...

+++ src/GameLayer/Components/World/Velocity.cpp
@@ -9,8 +9,8 @@ void Components::Velocity::normalize(float speed)
         return;
     }
     
-    float mag = velocity.getMagnitude();
-    if (mag == 0.0f) return;
+    double mag = velocity.getMagnitude();
+    if (mag == 0.0) return;
 
     velocity = velocity / mag * speed;
 }
